updates for preso and mod

+++ math.go
@@ -9,6 +9,6 @@ func subtract(num1, num2 float64) float64 {
 func multiply(num1, num2 float64) float64 {
 	return num1 * num2
 }
-func divide(num1, num2 int) float64 {
+func divide(num1, num2 float64) float64 {
 	return float64(num1 / num2)
 }
